api/requirements.txt
# Exousia API Dependencies
# FastAPI backend for YAML configuration management and GitHub integration

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# YAML Processing
pyyaml==6.0.1

# GitHub API
PyGithub==2.1.1
aiohttp==3.9.1

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Middleware & Utilities
python-dotenv==1.0.0
slowapi==0.1.9

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==23.12.1
ruff==0.1.11
